requirements.txt
mysql-connector-python>=8.0.26
pandas>=1.3.3
openpyxl>=3.0.9
xlrd>=2.0.1
python-dotenv>=0.19.1
argparse>=1.4.0
tabulate>=0.8.9
requests>=2.25.1
PyPDF2>=3.0.0
azure-storage-blob>=12.8.1
azure-storage-queue>=12.1.6
azure-storage-file-share>=12.4.2
azure-identity>=1.8.0
azure-mgmt-storage>=19.0.0
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7 